feature(fixingSound): now toggle works normally, needs some refactoring though
